dtc_server/requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.7
django-cors-headers==3.3.0
django-webpack-loader==0.7.0
djangorestframework==3.11.0
gunicorn==20.0.4
pytz==2020.1
sqlparse==0.3.1
static3==0.7.0
whitenoise==5.1.0
